requirements.txt
apache-airflow[slack]
